app/requirements.txt
asgiref==3.8.1
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
packaging==24.2
PyJWT==2.10.1
sqlparse==0.5.3
